+++ CHANGELOG.md
@@ -3,8 +3,9 @@
 * Add options locale, and useTeeOperator when using `setVerificationOptions` method.
 * Add option to use Reclaim's TEE+MPC Protocol for HTTP request claim verification & attestation.
 * Mention locale in headers for requests sent to reclaim sdk backend.
-* Add support for app links & deep links launch using `Reclaim.setAllowedAppLinks` API. This API can be used by custom injections
+* Add support for app links & deep links launch using `Reclaim.setAllowedAppLinks` API. This API can be used by provider user scripts.
 * Share the exact error message from backend on errors in `ReclaimSessionExpiredException`.
+* Add `Reclaim.updateUserAgent(userAgent:string)` API for updating user agent from provider user script.
 
 ## 0.24.0